Make it compile with Swift tools 5.8
The biggest change here is that Swift tools now support depending on a
system library directly rather than through another Swift package. This
removes the need for the CLibXML2 package, which was simply wrapping a
system library. Move the files from the package (a header file and a
modulemap) over to this one so that we can get rid of CLibXML2.